fix: move raw layer to an independent folder
