+++ plugins/helper/api_rawdata.go
@@ -1,4 +1,4 @@
-package core
+package helper
 
 import (
 	"fmt"
