+++ scripts/compile-plugins.sh
@@ -22,7 +22,7 @@ else
   # When rebuilding from all plugins, clean out old binaries first
   rm -rf bin/plugins/*
 
-  for PLUG in $(find $PLUGIN_SRC_DIR/* -maxdepth 0 -type d -not -name core -not -empty); do
+  for PLUG in $(find $PLUGIN_SRC_DIR/* -maxdepth 0 -type d -not -name core -not -name helper -not -empty); do
     NAME=$(basename $PLUG)
 
     echo "Building plugin $NAME to bin/plugins/$NAME/$NAME.so"
